Initial test on ubuntu machine

+++ customisations.sh
@@ -1,4 +1,7 @@
-## Ubuntu customisations: themes, icon sets...
-sudo apt-get install unity-tweak-tool numix-icon-theme-circle flattastic-suite ambiance-blackout-colorsmoka-icon-theme conky conky-all compizconfig-settings-manager compiz-plugins docky
+if DISTRIB | grep Ubuntu; then
+	## Ubuntu customisations: themes, icon sets...
+	sudo apt-get install unity-tweak-tool numix-icon-theme-circle flattastic-suite ambiance-blackout-colorsmoka-icon-theme conky conky-all compizconfig-settings-manager compiz-plugins docky
 
-sudo apt-get remove --purge unity-lens-friends unity-lens-shopping unity-scope-musicstores unity-scope-yelp unity-scope-zotero 
+	sudo apt-get remove --purge unity-lens-friends unity-lens-shopping unity-scope-musicstores unity-scope-yelp unity-scope-zotero 
+fi
+
